Clarity for Hare
Added a little bit explaining Hare would mean eps = 0 as it's technically just (votes/seats) with nothing additional; however, I left the eps in the code due to not wanting to mess with anything further in the code that may look for that variable when adjusting the quota variable during rounds. Also, leaving it in means more flexibility in quota options.